Refactor code structure and update requirements.txt
The code directory structure has undergone some modifications. In particular, the rag/retrieval/retrievers.py file has been renamed and shifted to a different location, i.e., retrievers/retrievers.py. Moreover, the requirements.txt file was updated in both rag/retrieval and the root directory, some packages were added while others were removed. Lastly, an email data CSV file (hilary-emails.csv) was added under the data/email_rag_evaluation_data directory. Additionally, renaming and updates in the
