+++ requirements.txt
@@ -1,8 +1,12 @@
 asyncio
 langchain>=0.1.6
+langchain-community
+langchain-openai
 langchain-experimental>=0.0.50
 openai>=1.12.0
 ragas>=0.1.0
 chromadb>=0.4.22
 python-dotenv
 pymupdf
+lark
+pandas
